KernelTensors

KernelTensors

TODO

Functions

Cloudy.KernelTensors.KernelTensor — Type.

KernelTensor{FT}

A kernel tensor approximation to a kernel function that can be used for different microphysical processes (e.g., coalescence, breakup, sedimentation).

source

Cloudy.KernelTensors.ConstantCoalescenceTensor — Type.

ConstantCoalescenceTensor{FT} <: KernelTensor

Represents a collision-coalescence tensor that is constant with respect to internal variables (e.g., mass or size, shape).

Constructors

ConstantCoalescenceTensor(c::Real)

Fields

source

Cloudy.KernelTensors.LinearCoalescenceTensor — Type.

LinearCoalescenceTensor{FT} <: KernelTensor

Corresponds to a collision-coalescence tensor that is a linear function with respect to internal variables (e.g., mass or size, shape).

Constructors

LinearCoalescenceTensor(c::Array{Real})

Fields

source