MassDistributions

Mass Distributions

TODO

Functions

Cloudy.MassDistributions.MassDistributionFunction — Type.

MassDistributionFunction{FT}

A particle mass distribution function, which can be initialized for various subtypes of assumed shapes in the microphysics parameterizations.

source

Cloudy.MassDistributions.Gamma — Type.

Gamma{FT} <: MassDistributionFunction

Represents particle mass distribution function of gamma shape.

Constructors

Gamma(n::Real, θ::Real, k::Real)

Fields

source

Cloudy.MassDistributions.Exponential — Type.

Exponential{FT}

Represents particle mass distribution function of exponential shape.

Constructors

Exponential(n::Real, θ::Real)

Fields

source

Cloudy.MassDistributions.compute_moment — Function.

compute_moment(dist, q)

dist - distribution of which the moment q is taken

q - is the potentially real-valued order of the moment

Returns the q-th moment of a particle mass distribution function.

source

Cloudy.MassDistributions.compute_density — Function.

compute_density(dist, x)

dist - is the particle mass distribution

x - is the point to evaluate the density of dist at

Returns the particle mass density evaluated at point x.

source

Cloudy.MassDistributions.update_params! — Function.

update_params(dist, mom)

dist - is a mass distirbution function

mom - is an array of moments

Updates the internal parameters of the mass distribution function dist.

source